% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buscar_unidad}
\alias{buscar_unidad}
\title{Validar y estandarizar el nombre de una unidad censal}
\usage{
buscar_unidad(unidad)
}
\arguments{
\item{unidad}{Nombre de la unidad (ej. "Lima", "Chachapoyas", "Peru").}
}
\value{
Una lista con:
\item{matched}{Boolean indicando si se encontró coincidencia.}
\item{tipo}{"nacional", "departamento", "provincia" o NULL.}
\item{nombre}{Nombre estandarizado de la unidad tal como figura en los datos.}
\item{departamento}{Si es provincia, el departamento al que pertenece.}
}
\description{
Validar y estandarizar el nombre de una unidad censal
}
