% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_educacion}
\alias{censo_educacion}
\title{Población censada por nivel educativo alcanzado (Censo 2025)}
\format{
Un objeto tibble con 10,750 filas y 7 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia.}
\item{area}{Área de residencia.}
\item{sexo}{Sexo de la población.}
\item{grupo_edad}{Grupos de edad de la población.}
\item{nivel_educativo}{Nivel alcanzado ("sin_nivel", "inicial", "primaria", "secundaria", etc.).}
\item{poblacion}{Cantidad de personas.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_educacion
}
\description{
Dataset ordenado con la población de 3 y más años de edad
según el máximo nivel educativo alcanzado.
}
\keyword{datasets}
