% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_estado_civil}
\alias{censo_estado_civil}
\title{Población censada por estado civil o conyugal (Censo 2025)}
\format{
Un objeto tibble con 1,134 filas y 7 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia ("Total").}
\item{area}{Área de residencia.}
\item{sexo}{Sexo de la población.}
\item{estado_civil}{Estado civil ("Soltero/a", "Casado/a", "Conviviente", etc.).}
\item{grupo_edad}{Grupos de edad de la población.}
\item{poblacion}{Cantidad de personas.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_estado_civil
}
\description{
Dataset ordenado con la población de 12 y más años de edad
según su estado civil o conyugal por grupos de edad.
}
\keyword{datasets}
