% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_etnicidad}
\alias{censo_etnicidad}
\title{Población censada por autoidentificación étnica (Censo 2025)}
\format{
Un objeto tibble con 2,268 filas y 7 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia.}
\item{area}{Área de residencia.}
\item{sexo}{Sexo de la población.}
\item{grupo_edad}{Grupos de edad.}
\item{grupo_etnico}{Etnia ("quechua", "aimara", "mestizo", "blanco", etc.).}
\item{poblacion}{Cantidad de personas.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_etnicidad
}
\description{
Dataset ordenado con la población que se identificó con algún grupo étnico.
}
\keyword{datasets}
