% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_hogar_numero}
\alias{censo_hogar_numero}
\title{Viviendas particulares por número de hogares (Censo 2025)}
\format{
Un objeto tibble con 2,808 filas y 7 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia.}
\item{area}{Área de residencia.}
\item{tipo_vivienda}{Tipo de vivienda.}
\item{indicador}{"Viviendas particulares" o "Ocupantes presentes".}
\item{numero_hogares}{Número de hogares ("hogares_1", "hogares_2", etc.).}
\item{valor}{Cantidad.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_hogar_numero
}
\description{
Distribución de viviendas particulares ocupadas según la cantidad de hogares
que albergan en su interior.
}
\keyword{datasets}
