% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_seguro_salud}
\alias{censo_seguro_salud}
\title{Población censada por tipo de seguro de salud (Censo 2025)}
\format{
Un objeto tibble con 1,323 filas y 7 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia.}
\item{area}{Área de residencia.}
\item{sexo}{Sexo de la población.}
\item{grupo_edad}{Grupos de edad.}
\item{tipo_seguro}{Seguro ("sis", "essalud", "ffaa_policiales", "privado", "otro", "ninguno", "total").}
\item{poblacion}{Cantidad de personas.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_seguro_salud
}
\description{
Dataset ordenado con la afiliación a algún tipo de seguro de salud.
}
\keyword{datasets}
