% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censo_vivienda}
\alias{censo_vivienda}
\title{Viviendas particulares por condición de ocupación (Censo 2025)}
\format{
Un objeto tibble con 1,890 filas y 6 columnas:
\describe{
\item{departamento}{Nombre del departamento.}
\item{provincia}{Nombre de la provincia.}
\item{area}{Área de residencia.}
\item{tipo_vivienda}{Tipo de vivienda (Casa independiente, departamento, etc.).}
\item{condicion_ocupacion}{Condición ("ocupada_total", "desocupada_total", etc.).}
\item{valor}{Cantidad de viviendas.}
}
}
\source{
Instituto Nacional de Estadística e Informática (INEI) - Censos Nacionales 2025.
}
\usage{
censo_vivienda
}
\description{
Dataset ordenado de viviendas particulares según su tipo y
condición de ocupación.
}
\keyword{datasets}
