% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limpiar_texto}
\alias{limpiar_texto}
\title{Limpiar texto para comparaciones insensibles a mayúsculas y acentos}
\usage{
limpiar_texto(x)
}
\arguments{
\item{x}{Vector de caracteres.}
}
\value{
Vector de caracteres limpio, en mayúsculas, sin acentos y recortado.
}
\description{
Limpiar texto para comparaciones insensibles a mayúsculas y acentos
}
\examples{
limpiar_texto("Áncash ")
}
