% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listar_departamentos}
\alias{listar_departamentos}
\title{Obtener lista única de departamentos disponibles}
\usage{
listar_departamentos()
}
\value{
Vector con los nombres de departamentos.
}
\description{
Obtener lista única de departamentos disponibles
}
