% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listar_provincias}
\alias{listar_provincias}
\title{Obtener lista única de provincias disponibles}
\usage{
listar_provincias(depto = NULL)
}
\arguments{
\item{depto}{Opcional. Filtrar provincias por departamento.}
}
\value{
Vector con los nombres de provincias.
}
\description{
Obtener lista única de provincias disponibles
}
