% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{plot_educacion}
\alias{plot_educacion}
\title{Graficar la distribución del nivel educativo alcanzado}
\usage{
plot_educacion(unidad)
}
\arguments{
\item{unidad}{Nombre de la unidad censal.}
}
\value{
Un objeto ggplot2.
}
\description{
Graficar la distribución del nivel educativo alcanzado
}
