% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{plot_etnicidad}
\alias{plot_etnicidad}
\title{Graficar la distribución de autoidentificación étnica}
\usage{
plot_etnicidad(unidad)
}
\arguments{
\item{unidad}{Nombre de la unidad censal.}
}
\value{
Un objeto ggplot2.
}
\description{
Graficar la distribución de autoidentificación étnica
}
