% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{plot_piramide_poblacion}
\alias{plot_piramide_poblacion}
\title{Graficar la pirámide de población de una unidad censal}
\usage{
plot_piramide_poblacion(unidad)
}
\arguments{
\item{unidad}{Nombre de la unidad (ej. "AMAZONAS", "CHACHAPOYAS" o "PERÚ").}
}
\value{
Un objeto ggplot2.
}
\description{
Graficar la pirámide de población de una unidad censal
}
