% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{plot_servicios_agua}
\alias{plot_servicios_agua}
\title{Graficar la distribución de viviendas por procedencia de agua}
\usage{
plot_servicios_agua(unidad)
}
\arguments{
\item{unidad}{Nombre de la unidad censal.}
}
\value{
Un objeto ggplot2.
}
\description{
Graficar la distribución de viviendas por procedencia de agua
}
